local_ELT_prefect/ELT_requirements.txt
pandas==2.1.4
pyarrow==14.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
prefect==2.14.9
prefect-dbt==0.4.1
dbt-postgres==1.7.6
